backend/requirements.txt
asgiref==3.2.7
coverage==5.1
Django==3.0.5
django-cors-headers==3.3.0
django-extensions==2.2.9
django-filter==2.2.0
djangorestframework==3.10.3
djangorestframework-simplejwt==4.4.0
Faker==0.9.1
html5lib==1.0.1
mixer==6.1.3
Pillow==7.1.2
psycopg2==2.8.4
PyJWT==1.7.1
PyPDF2==1.26.0
python-dateutil==2.8.1
pytz==2020.1
PyYAML==5.3.1
reportlab==3.5.42
six==1.15.0
sqlparse==0.3.1
text-unidecode==1.2
uritemplate==3.0.1
webencodings==0.5.1
xhtml2pdf==0.2.4
